Database/Hibernate/Inheritence/src/main/java/ru/yaal/examples/database/hibernate/inheritance/single/discriminator/CarSingleDiscriminator.java
package ru.yaal.examples.database.hibernate.inheritance.single.discriminator;

import javax.persistence.DiscriminatorColumn;
import javax.persistence.DiscriminatorType;
import javax.persistence.DiscriminatorValue;
import javax.persistence.Entity;
import javax.persistence.Table;

@Entity
@Table(name = "car_single_discriminator")
@DiscriminatorColumn(name = "dtype", discriminatorType = DiscriminatorType.STRING, length = 31)
@DiscriminatorValue("car_signle_discriminator")
public class CarSingleDiscriminator extends TransportSingleDiscriminator {
    private String fuel;

    public String getFuel() {
        return fuel;
    }

    public void setFuel(String fuel) {
        this.fuel = fuel;
    }
}
